requirements.txt
gitpython
reportlab
tqdm
